pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "canctool"
version = "1.0.0"
description = "LLM Tool Call Wrapper Service with multi-provider support"
readme = "README.md"
requires-python = ">=3.8.1"
license = {text = "MIT"}
authors = [
    {name = "CancTool Team", email = "support@canctool.com"}
]
keywords = ["llm", "openai", "tool-calling", "api", "wrapper", "multi-provider"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-asyncio>=0.20.0",
    "black>=22.0.0",
    "isort>=5.10.0"
]

[project.urls]
Homepage = "https://github.com/canctool/canctool"
Documentation = "https://canctool.readthedocs.io"
Repository = "https://github.com/canctool/canctool.git"
"Bug Tracker" = "https://github.com/canctool/canctool/issues"

[project.scripts]
canctool = "main:main"
canctool-server = "main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["canctool*"]
exclude = ["tests*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100
